SS-9745: Add evals test
