+++ features/webmcp/evals/__tests__/webmcpEvals.test.ts
@@ -1,0 +1,14 @@
+import {
+	loadWebmcpEvalScenarios,
+	runWebmcpEvalScenario,
+} from "../runWebmcpEvalScenarios";
+
+describe("webmcp eval scenarios", () => {
+	const scenarios = loadWebmcpEvalScenarios();
+
+	it.each(scenarios)("$id: $description", async (scenario) => {
+		const failure = await runWebmcpEvalScenario(scenario);
+
+		expect(failure).toBeNull();
+	});
+});
